==> data/4-Reconstitution_CoordGéo_GPH-IREL.ipynb <==
<h4>Lily Grumbach
    <br>M1 Humanités numériques - Université PSL</h4>
<h1><center>Rendu TAIS-TAL </center></h1>
<h2><center>Partie TAIS</center></h2>
<h3><center>4/5 : Reconstituer les coordonnées géographiques des GPH et de l'IREL</center></h3>

<h3><u>PLAN:</u></h3>

1) Harmonisation des données Géopolitiques pour le fond de carte de 1914
<br>
    
2) Explorer les bases de données à ma disposition (IREL et GPH) et les préparer pour l'extraction des coordonnées géographiques 
<br/>
 
3) Reconstituer les liens URL de la base de données IREL des Archives nationales d'Outre-mer.

<h4>4) Reconstituer les coordonnées géographiques des GPH et de l'IREL</h4>
 
5) Géoréférencer les données de mes revues


<b>Stratégie en place :  </b>

Nous avons eu recours à trois sources de coordonnées géographiques: 
* **Wikidata** indiqués par le medialab 
* l'**IREL**, la base de données géographique de l'inventaire en ligne des Archives nationales d'Outre Mer
* **Geonames**, pour le meilleur et pour le pire.

<h3>Objectifs de la partie : </h3>

Récupérer les coordonnées décimales d'entités géographiques historiques pour pouvoir l'appliquer sans problème à l'avenir.

#module
import desambiguisation

##General : 
from desambiguisation import ListEntities2df
##GPH
from desambiguisation import MatchGPELOC_IREL,Extract_longlat_WikiData
##IREL : 
from desambiguisation import nettoyage_df_IREL,urlencode,Extract_longlat_IREL,IREL_Nettoyage_AdminLieuDit,nettoyage_desambiguisation

#Généraux
import pandas as pd
import numpy as np
import re

#Webscrapping
from bs4 import BeautifulSoup
import geocoder
from lxml import html
import requests

#Wikidata 
import qwikidata
from qwikidata.entity import WikidataItem, WikidataLexeme, WikidataProperty
from qwikidata.linked_data_interface import get_entity_dict_from_api

# 0. Import des données : 

#Liste des entités GPELOC à annoter :
df_GPELOC=pd.read_csv("./output/2-Exploration_GPH-IREL/GPELOC_à_annoter.csv")

#df des lieux en 1914 avec leur lien Wikidata :
df_WikiQGIS = pd.read_csv("./output/2-Exploration_GPH-IREL/WikiQGIS.csv")

#df de l'ensemble des lieux de GPH avec leur lien wikidata
df_wikiData = pd.read_csv("./GeoPolHist-202103/medialab-GeoPolHist-fb19b66/data/GeoPolHist_entities.csv")
df_wikiData=df_wikiData.rename(columns={"GPH_name":"NAME"})

#df des lieux dans l'IREL avec leurs coordonnées géographiques


df_GPELOC

dico_IREL ={}
n=0

with open('./IREL/liste_lieux_IREL.js','r') as f:
    lignes = f.readlines()
    
    for l in lignes:
        l=l.split("\n")
        dico_IREL[n]=l
        n+=1

df_IREL = nettoyage_df_IREL(dico_IREL)

df_IREL=df_IREL.rename(columns={"Lieu-dit":"NAME"})
# df_IREL

# 1. GPH : Retrouver dans wikidata les coordonnées géographiques

=> La méthode que nous avons mis en place pour récupérer l'ensemble des coordonnées n'a pas été concluante malgré les nombreux efforts pour éviter l'ensemble des messages d'erreur provenantde WikiData. Nous envisageons reprendre la fonction du module pour l'appliquer par silots de 100 lieux afin d'avoir plus de contrôle dessus.

#### Pour extraire TOUTES les coordonnées géographiques des entités de `GPH`

pattern = r"(Q\d+)"
df_wikiData["Q_code"] = df_wikiData["wikidata"].str.extract(pattern)
df_wikiData.to_csv("./output/4-Reconstitution_CoordGéo_GPH-IREL/GPH_Wikidata_Qcode.csv")

List_Q_code_all =df_wikiData["Q_code"].tolist()
output_path_all = "./output/4-Reconstitution_CoordGéo_GPH-IREL/GPELOC_GPH_coordonnees_all.csv"

## ETAPE TRES LONGUE. Nous avons enregistré le resultat dans output_path_all 
# Extract_longlat_WikiData(List_Q_code_all,output_path_all)


#Importation des donnéesavec Q_code, latitude et longitude 
df_WikiData_longlat_all = pd.read_csv(output_path_all)

df_wikiData

Merge du df contenant les long/lat avec le df contenant les Noms et GPH_code de chaque entité

df_WikiData_longlat_Name = pd.merge(df_wikiData[["GPH_code","NAME","Q_code"]],df_WikiData_longlat_all,how="inner",on="Q_code")

Faire une estimation du nombre de Q_code n'ayant pas de coordonnées dans dans WikiData : 

print("Nombre de coordonnées manquantes dans WikiData : ",len(df_WikiData_longlat_Name.loc[df_WikiData_longlat_Name["latitude"]==0]))

#### Récupérer les long/lat des données  `GPELOC` de 1898 à 1908 et les merge avec les longitudes et latitudes

La majeur différence entre la liste GPELOC à notre disposition et la base de donnée de GPH est la langue. GPH est en anglais tandis que GPELOC en français. On va donc ici d'abord "traduire" les entités nommées de la liste GPELOC en anglais.

N'ayant pas réussi à faire fonctionner `googletrans ` nous avons traduit directement sur Deepl en ligne

with open ("./desambiguisation/list_GPELOC_anglais.txt") as f:
    list_GPELOC_anglais = (f.readlines())

for i in range (len(list_GPELOC_anglais)):
    list_GPELOC_anglais[i]=list_GPELOC_anglais[i].replace(",\n","")
list_GPELOC_anglais

df_GPELOC_anglais=pd.DataFrame({"NAME":list_GPELOC_anglais})

##Faire un dictionnaire:
list_GPELOC=df_GPELOC.NAME.tolist()
dico_placeName = {}
for i in range(len(list_GPELOC_anglais)):
    placeNameEnglish = list_GPELOC_anglais[i]
    placeNameFrançais = list_GPELOC[i]
    dico_placeName[placeNameEnglish] = placeNameFrançais  
# dico_placeName

EN ANGLAIS : Création d'un df des entités de GPELOC trouvées dans WikiData_longlat

df_WikiData_longlat_18981914_anglais =pd.merge(df_WikiData_longlat_Name,df_GPELOC_anglais,how="inner",on="NAME")

#On remarque qu'il y a un doublon : 
df_WikiData_longlat_18981914_anglais=df_WikiData_longlat_18981914_anglais.drop_duplicates(keep="first")

df_WikiData_longlat_18981914_anglais.to_csv("./output/4-Reconstitution_CoordGéo_GPH-IREL/GPH_GPELOC_longlat_anglais.csv",index=False)

df_WikiData_longlat_18981914_anglais

TRADUCTION du nom des lieux en français pour pouvoir en suite merge avec DF_GPELOC : 

df_WikiData_longlat_18981914=df_WikiData_longlat_18981914_anglais

#Méthode barbare, jesais:
liste_placeName_geoloc=[]
for placeNameEnglish in  df_WikiData_longlat_18981914_anglais["NAME"]:
    placeNameFrançais = dico_placeName[placeNameEnglish]
    liste_placeName_geoloc.append(placeNameFrançais)

df_WikiData_longlat_18981914["NAME"]=liste_placeName_geoloc
df_WikiData_longlat_18981914

#### Cette démarche a été concluante puisque l'on a récupéré 40 entités au lieu des 23 déclarées dans la partie 2 de ce rendu !

### On merge la liste des latitudes/longitudes `GPELOC_GPH ` (càd trouvées dans la base de données GPH) avec la liste de l'ensemble  des lieux à géoréférencer pour nos revues, `df_GPELOC`

# list_placesName_18981914 = df_WikiData_longlat_18981914.NAME.tolist()
# for placeName in list_placesName_18981914 :
#     list_GPELOC.remove(placeName)

On a alors la liste ` list_GPELOC ` qui ne contient que les noms de lieux dont il reste à trouver leslong/lat

# 2. IREL : Retrouver par webscrapping les coordonnées géographiques des Archives d'Outre-mer

Beaucoup plus compliqué.....

df_IREL

## 2.1. Chercher quels noms de lieux présents dans ma revue se trouve dans celle des IREL  

#### Import du df de ce que l'on peut et ne peut pas trouver avec l'IREL

comparaison_matchIREL=pd.read_csv("./desambiguisation/comparaison_match_IREL.csv")
comparaison_matchIREL

#### Import du df IREL avec les coordonnées

df_IREL_url = pd.read_csv("./output/3-Reconstitution_URL_IREL/IREL_ens_url.csv")
df_IREL_url

#### Maintenant on cherche à extraire de ` df_IREL ` les coordonnées des lieux qui ont matché 

Le principal problème est que pour chaque entité qui "match", on peut avoir de nombreux résultats.Par exemple, "Algérie" compte plus 

df_ensemble_byPlaceName=pd.read_csv("./output/3-Reconstitution_URL_IREL/IREL_parlieuxreconnus.csv")


df_ensemble_byPlaceName

liste_urlIREL = list(df_ensemble_byPlaceName["IREL_url"].str.split("http",expand=True)[2])
for i in range (len(liste_urlIREL)):
    liste_urlIREL[i]="http"+str(liste_urlIREL[i])
# "http"+liste_urlIREL[3]
liste_urlIREL

# À activer la 1er fois : 
df_ensemble_byPlaceName["IREL_url"]=liste_urlIREL

df_ensemble_byPlaceName

df_IREL

df_ensemble_byPlaceName2 = pd.merge(df_ensemble_byPlaceName,df_IREL_url,how='left',on="IREL_url")
df_ensemble_byPlaceName2 = df_ensemble_byPlaceName2[["IREL_url","NAME","latitude","longitude"]]


df_ensemble_byPlaceName2

## Merge le df des longitudes/latitudes avec df_GPELOC

df_GPELOC

df_ensemble_byPlaceName2.to_csv("./output/4-Reconstitution_CoordGéo_GPH-IREL/IREL_GPELOC")

# 3. Concaténer Les long/lat de IREL, GPH et les entités qu'il reste à anoter : 

df_GPH_IREL_coord=pd.concat([df_WikiData_longlat_18981914,df_ensemble_byPlaceName2])
df_GPH_IREL_coord

df_GPH_IREL_GPELOC_coord = pd.merge(df_GPH_IREL_coord,df_GPELOC,how='right',on='NAME')

df_GPH_IREL_GPELOC_coord.to_csv("./output_finaux/COORD_GPH_IREL_GPELOC.csv")
df_GPH_IREL_GPELOC_coord.to_csv("./output/4-Reconstitution_CoordGéo_GPH-IREL/COORD_GPH_IREL_GPELOC.csv")

df_GPH_IREL_GPELOC_coord

df_GPH_IREL_GPELOC_coord
